Fixed bug in gridsearch and calculation of pagerank ROC/PR curves.
During the grid search, the number of predicted genes was logged falsely. This number is
sometimes important to get a glimpse of how many genes are thought to be involved.
Furthermore, the dataframe from the pagerank algorithm was not in the correct order and
therefore, the PR and ROC curves were wrong. This was fixed now as well.