Renamed GCN to EMOGI and verified notebooks work.
The GCN directory contained the EMOGI code as well as all notebooks.
Now, notebooks are located in the analysis folder and the GCN folder was
renamed to EMOGI.
Further, I verified that all notebooks basically work in my environment
properly.
Each notebook has paths set in the beginning so that it should be
possible to adapt easily to a different environment with these data
sets.