Fixed but and improved perturbation experiments.
In detail, the edge swaps are adapted to approximately produce the
correct percentages of swaps (25%, 50%, 75%).

A bug resulted in higher node degree of cancer genes. This was because
they were more likely to be first in the edge list files of CPDB because
they have generally more interactions. This resulted in them being first
in the random network generation where we use a preferential-attachment
model.
The bug has been fixed but that was annoying and responsible for the
non-random performance of an all-shuffled model.